chore: sync repo config
