+++ .nvmrc
@@ -1,1 +1,1 @@
-22.12.0
+24.4.1
