+++ examples/react/tanstack-start/package.json
@@ -18,7 +18,7 @@
     "vinxi": "^0.5.7"
   },
   "devDependencies": {
-    "@types/node": "^22.10.6",
+    "@types/node": "^24.1.0",
     "@types/react": "^19.0.7",
     "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react": "^4.5.2",
