+++ scripts/publish.ts
