+++ scripts/verify-links.ts
@@ -1,13 +1,12 @@
 import { existsSync, readFileSync, statSync } from 'node:fs'
 import path, { resolve } from 'node:path'
-import fg from 'fast-glob'
-// @ts-ignore
+import { glob } from 'tinyglobby'
+// @ts-ignore Could not find a declaration file for module 'markdown-link-extractor'.
 import markdownLinkExtractor from 'markdown-link-extractor'
 
 function isRelativeLink(link: string) {
   return (
     link &&
-    !link.startsWith('/') &&
     !link.startsWith('http://') &&
     !link.startsWith('https://') &&
     !link.startsWith('//') &&
@@ -87,7 +86,7 @@ function fileExistsForLink(
 
 async function findMarkdownLinks() {
   // Find all markdown files in docs directory
-  const markdownFiles = await fg('docs/**/*.md', {
+  const markdownFiles = await glob('docs/**/*.md', {
     ignore: ['**/node_modules/**'],
   })
 
@@ -110,7 +109,7 @@ async function findMarkdownLinks() {
     })
 
     if (filteredLinks.length > 0) {
-      filteredLinks.forEach((link: any) => {
+      filteredLinks.forEach((link) => {
         const href = typeof link === 'string' ? link : link.href
         fileExistsForLink(href, file, errors)
       })
